bugfix rake task

+++ app/models/user_mail.rb
@@ -38,19 +38,19 @@ class UserMail < ApplicationRecord
   private
   def upload_location
     upload_path = self.upload_path || ''
+
+    file_path = self.source_file.path unless self.source_file.file.nil?
+
     if self.source_file.present?
-      file_path = self.source_file.path
-      if upload_path != file_path
-        if File.exists?(upload_path)
-          File.rename(upload_path, file_path)
-        elsif File.exists?(file_path)
-          self.upload_path = file_path
-        else
-          Rails.logger.error("File for attachment #{id} cannot be found")
-          # TODO: rescan mail?
-        end
-        self.upload_path = file_path
-      end
+      self.upload_path=file_path
+    elsif File.exists?(upload_path) && file_path
+      File.rename(upload_path, file_path)
+      self.upload_path = file_path
+    else
+      Rails.logger.error("File for attachment #{id} cannot be found")
+      # TODO: rescan mail?
     end
+
   end
+
 end
